Check the crown file compatibility

Check the crown file compatibility with the managecrownsdata functions. Return text, that give informations about the file modifications needed.

Usage

check_crownsFile(path_crowns)

Arguments

path_crowns

chr. Path to the crown file

Value

Text that give you information about your file. Indicates whether your file will be compatible or not for the other functions of the package. Pay attention to the line starting with '– ERROR –'.