Create xlsx labeling file

A function to create the xlsx file to do the labeling

Usage

create_labelingFile(
  crownsFile,
  site = NULL,
  dates = NULL,
  save_xlsx = FALSE,
  directory = NULL
)

Arguments

crownsFile

sf object. The crowns delineation file. It should include 'id', 'species', 'genus' and 'family' columns.

site

chr. site name.

dates

chr. vector of dates (format should be 'YYYY_MM_DD', p.e '20220925').

save_xlsx

logical. If TRUE, it will save the table as xlsx file. Indicate the path as the directory parameters

directory

The path where to save the xlsx file.

Examples

library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(stringr)

rgb_paths <- list.files(
file.path(
system.file(package="managecrownsdata"), 'rgb/'),
full.names = TRUE
)
path_crownsFile <- file.path(
system.file(package="managecrownsdata"),
'crowns/Bouamir_crowns.gpkg'
)
crownsFile <- sf::read_sf(path_crownsFile)

site = 'Bouamir'
dates <- paste(
str_sub(stringr::str_split(basename(rgb_paths), '_', simplify = TRUE)[,2],1,4),
str_sub(stringr::str_split(basename(rgb_paths), '_', simplify = TRUE)[,2],5,6),
str_sub(stringr::str_split(basename(rgb_paths), '_', simplify = TRUE)[,2],7,8),
sep = ''
)

check_crownsFile(crownsFile = crownsFile)
#> ##########     VARIABLES CHECK     ##########
#> -                                           -
#> -- ERROR --  :   geometry variable missing or not well named
#> --- OK ----  :   id
#> --- OK ----  :   family
#> --- OK ----  :   genus
#> --- OK ----  :   species
#> --- OK ----  :   plot_name
#> --- OK ----  :   code_sp
#> -                                           -
#> -                                           -
#> ##########           CRS           ##########
#> -                                           -
#> CRS : WGS 84 / UTM zone 33N
#> -                                           -
#> -                                           -
#> ##########   DUPLICATED ID CHECK   ##########
#> -                                           -
#> --- OK ----  :  There is no duplicated id

crownsFile <- crownsFile %>% rename(
geometry = geom
)

check_crownsFile(crownsFile = crownsFile)
#> ##########     VARIABLES CHECK     ##########
#> -                                           -
#> --- OK ----  :   geometry
#> --- OK ----  :   id
#> --- OK ----  :   family
#> --- OK ----  :   genus
#> --- OK ----  :   species
#> --- OK ----  :   plot_name
#> --- OK ----  :   code_sp
#> -                                           -
#> -                                           -
#> ##########           CRS           ##########
#> -                                           -
#> CRS : WGS 84 / UTM zone 33N
#> -                                           -
#> -                                           -
#> ##########   DUPLICATED ID CHECK   ##########
#> -                                           -
#> --- OK ----  :  There is no duplicated id

create_labelingFile(
crownsFile = crownsFile,
site = site,
dates = dates,
save_xlsx = FALSE
)
#> # A tibble: 41 × 14
#>    site       id species   genus family     n obs   update `20220427` `20220511`
#>    <chr>   <dbl> <chr>     <chr> <chr>  <int> <lgl> <lgl>  <lgl>      <lgl>     
#>  1 Bouamir   665 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#>  2 Bouamir   666 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#>  3 Bouamir   668 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#>  4 Bouamir   669 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#>  5 Bouamir   670 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#>  6 Bouamir   671 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#>  7 Bouamir   673 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#>  8 Bouamir   674 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#>  9 Bouamir   676 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#> 10 Bouamir   678 Desborde… Desb… Irvin…    19 NA    NA     NA         NA        
#> # ℹ 31 more rows
#> # ℹ 4 more variables: `20220525` <lgl>, `20220610` <lgl>, `20220624` <lgl>,
#> #   Comm <lgl>