Extract and export the non NA area of images.

This function extract the area of images covered by non NA values and return an sf object with one polygon per image.

Usage

extract_bboxImages(
  path_in,
  crs = NULL,
  dates = NULL,
  directory = NULL,
  filename = NULL
)

Arguments

path_in

vector with the full paths to the RGB rasters images.

crs

crs. Object of class 'crs', could be get from st_crs(..). If NULL, it will use and transform all the data into the crs of the first RGB image.

dates

chr. Vector with dates (format should be '%Y_%m_%d', p.e '2022_09_25'). The order of the dates should match with the order of the dates of the image in the path_in

directory

chr. The path to the directory used to export the vectors. One file will be export per date. If NULL, the vector will not be exported.

filename

logical. The name of the files.

Examples

rgb_paths <- list.files(
file.path(
system.file(package="managecrownsdata"), 'rgb/'),
full.names = TRUE
)

#bbox <- extract_bboxImages(
#path_in = rgb_paths,
#directory = my_directory)