Extract RGB metrics

The function extracts RGB metrics (red, green, blue, sumrgb, rcc, gcc, bcc, gndvi, gli) for each crown at each date. It first compute the metrics for the whole mosaics and then extracts the value at the crown scale using the exactextractr::exact_extract function. The mean and / or the variance can be extracted (see 'fun' parameter).

Usage

extract_rgbValues(
  crownFile,
  RGB_paths,
  site = NULL,
  date = NULL,
  crs = NULL,
  fun = "all",
  infos = FALSE
)

Arguments

crownFile

A sf object with the crowns delineation.

RGB_paths

a list with the full paths to the RGB rasters.

site

chr. name of the site, p.e 'Mbalmayo'.

date

chr. vector of dates (format should be 'yyyy_mm_dd', p.e '2022_09_25'). The order of the dates should match with the order of the RGB_paths !

crs

crs. Object of class 'crs', could be get from st_crs(..). If NULL, it will use and transform all the data into the crs of the first RGB image.

fun

chr. Specify the function used in the 'fun' parameter of the exactextractr::exact_extract function to extract RGB values, it could be 'mean', 'var' or 'all'.

infos

logical. Specify whether or not to return details of the extraction. When TRUE, specify the crowns which has not been extracting per date, because they were out of the image.

Examples

if (FALSE) { # \dontrun{
library(sf)
library(dplyr)
library(stringr)

path <- system.file(package="phenobsappli")
RGB_paths <- list.files(file.path(path, 'tif'), full.names = TRUE)
crowns <- read_sf (list.files(file.path(path, 'shp'), full.names = TRUE))
date = sapply( str_split( basename( RGB_paths ),'_' ),
                     function(x) str_remove(paste(x[2], x[3], x[4], sep ='_'),'.tif') )
crs = sf::st_crs(stars::read_stars(RGB_paths[1], proxy = TRUE))
site = 'Mbalmayo'

test <- extract_rgbValues(
   crowns,
   RGB_paths,
   crs = crs,
   site = site,
   date = date,
   fun = 'all',
   infos = TRUE)
} # }