Format labels values from wide data to long data

A function to format labels data from wide to long format

Usage

pivot_Labels(wideLabels, simplify_labels = FALSE)

Arguments

wideLabels

tibble or dataframe which contains labeling data directly import from the xlsx file.

simplify_labels

logical Decompose and simplify the labels when TRUE. By defaut it is FALSE.

Value

tibble

Examples

library(readxl)

raw_labels <- read_excel(
file.path(
system.file(package="managecrownsdata"),
'xlsx/labeling_file_Bouamir.xlsx')
)

pivot_Labels(raw_labels)
#> # A tibble: 2,050 × 12
#>    site       id species    genus family     n date       phenophase obs   Comm 
#>    <chr>   <int> <chr>      <chr> <chr>  <dbl> <date>     <chr>      <chr> <chr>
#>  1 Bouamir   665 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#>  2 Bouamir   666 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#>  3 Bouamir   668 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#>  4 Bouamir   669 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#>  5 Bouamir   670 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#>  6 Bouamir   671 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#>  7 Bouamir   673 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#>  8 Bouamir   674 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#>  9 Bouamir   676 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#> 10 Bouamir   678 Desbordes… Desb… Irvin…   248 2022-03-26 L          YN    NA   
#> # ℹ 2,040 more rows
#> # ℹ 2 more variables: update <chr>, Usable_crown <lgl>

pivot_Labels(raw_labels, simplify_labels = TRUE)
#> # A tibble: 2,050 × 21
#>    site       id species      genus family     n date       phenophase PPfoliar1
#>    <chr>   <int> <chr>        <chr> <chr>  <dbl> <date>     <chr>      <chr>    
#>  1 Bouamir   665 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#>  2 Bouamir   666 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#>  3 Bouamir   668 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#>  4 Bouamir   669 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#>  5 Bouamir   670 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#>  6 Bouamir   671 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#>  7 Bouamir   673 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#>  8 Bouamir   674 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#>  9 Bouamir   676 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#> 10 Bouamir   678 Desbordesia… Desb… Irvin…   248 2022-03-26 L          L        
#> # ℹ 2,040 more rows
#> # ℹ 12 more variables: PPfoliar2 <chr>, PPFlo <dbl>, PPFr <dbl>,
#> #   PPFlo_uncertainty <dbl>, PPFr_uncertainty <dbl>, desynchr <dbl>,
#> #   PPfoliar1_uncertainty <dbl>, PPfoliar2_uncertainty <dbl>, obs <chr>,
#> #   Comm <chr>, update <chr>, Usable_crown <lgl>