Extract crowns images

The function extracts and save .jpeg images for each crown at each date.

Usage

extract_crownsImages(
  path_images,
  crownsFile,
  path_bbox,
  path_out,
  site = NULL,
  dates = NULL,
  crs = NULL,
  parallel = FALSE,
  N_cores = NULL,
  update = FALSE,
  specific_quality = FALSE,
  width = 250,
  height = 300
)

Arguments

path_images

A sf object for the crowns with an 'id' variable.

crownsFile

a list with the full paths to the RGB rasters.

path_bbox

The path to the non NA Bbox return by the function extract_bboxImages()

path_out

chr. The path to the directory use to stored the images. The function will create the folder, It doesn't need to exists.

site

chr. name of the site, p.e 'Mbalmayo'.

dates

chr. Vector with dates (format should be '%Y_%m_%d', p.e '2022_09_25'). The order of the dates should match with the order of the dates of the image in the path_images

crs

crs. Object of class 'crs', could be get from st_crs(..). If NULL, it will use and transform all the data into the crs of the first RGB image.

parallel

xx

N_cores

xx

update

xx

specific_quality

Logical, if TRUE the quality of image will be determine by the height and width arguments.

width

if specific_quality = TRUE, the width of the device

height

if specific_quality = TRUE, the height of the device

Details

The extract_crownsImages() create one folder per id and save the images. The folder names are 'crown_the idthe specie name' for exemple 'crown_5_Lophira alata'. The the images names are 'crownthe idthe specie namethe date.jpeg' for exemple 'crown_5_Lophira alata_2022-11-08.jpeg'. The function upload square image with neighbouring tree and the title is add at the top, image size is 720*825 pixels. When specific_quality is TRUE, the image size can be changed by specifying height and width parameters.